Added profile and upgrade pages responsive UI in light mode
No functionality added yet. ChessBoard was used for the image of the board just as was done in the homepage. ChessBoard was
updated to take in color scheme as props and user is given the ability to specify both light and dark colors. Potential improvement
could be to have a component or extension just for when we need the board image to avoid actually constructing a chess object and board.
Spartan font was added to match design.