Implemented updating profile and user context & auth session to get user id
In order to make requests using the user id, the session callback and
interface were updated such that the user id would be included in the session.

Changed session access in user api from getSession to unstable_getServerSession,
and added a PUT endpoint for profile updates.

Updated Profile to be able to make updates and added a UserInfoContext callback
to be able to update the state from any child and automatically rerender
dependant components.